+++ NEWS.md
@@ -1,20 +1,28 @@
-wallace 0.0.4
+bioclima 0.1.0
+=============
+- Easier names for internal functions
+- Option to add bioclimatic variables bio24-35
+- bionames() to get names easily
+- Option to add static periods. Useful when you wnat to use same periods in time-series
+- New stats_clima() to make your own summaries of variables (e.g., evapotraspiraion, cloudiness)
+
+bioclima 0.0.4
 =============
 - Fix bug in mismatch_NA().
 - Some small changes in documenting the package.
 
-wallace 0.0.3
+bioclima 0.0.3
 =============
 - Changing pipe and adding call to terra functions and internal functions. Thanks to the contribution of [@francisvolh](https://github.com/francisvolh)
 - Updated node in pkgdown.
 
-wallace 0.0.2
+bioclima 0.0.2
 =============
 - Vignette in Spanish to be show in the "Modelamiento de Nichos Ecológicos" facebook group
 - Add stdev for population in bio04 and bio15
 - No round of variables
 - More infomation in README file
 
-wallace 0.0.1
+bioclima 0.0.1
 =============
 - First code public